+++ selfdrive/ui/layouts/settings/driving_mannerisms.py
@@ -10,6 +10,8 @@ from openpilot.selfdrive.ui.layouts.settings.nap_content import (
   HYPERMILE_DESCRIPTION,
   HYPERMILE_HILL_CLIMB_DESCRIPTION,
   HYPERMILE_STEP_DOWN_DESCRIPTION,
+  MAP_SPEED_ACCEL, MAP_SPEED_ACCEL_DEFAULT, MAP_SPEED_ACCEL_DESCRIPTION,
+  MAP_SPEED_ACCEL_LABELS,
   NAP_DRIVER_LAT_HANDOFF,
   NAP_HYPERMILE,
   NAP_HYPERMILE_HILL_CLIMB,
@@ -69,6 +71,17 @@ class DrivingMannerismsLayout(Widget):
     )
     self._all_items.append(self._adaptive_accel)
 
+    accel = int(self._params.get("NAPMapSpeedAccel", return_default=True) or MAP_SPEED_ACCEL_DEFAULT)
+    self._accel_buttons = multiple_button_item(
+      "Acceleration",
+      MAP_SPEED_ACCEL_DESCRIPTION,
+      buttons=MAP_SPEED_ACCEL_LABELS,
+      button_width=72,
+      selected_index=self._accel_index(accel),
+      callback=self._on_accel,
+    )
+    self._all_items.append(self._accel_buttons)
+
     follow_dist = self._params.get(NAPParamKeys.FOLLOW_DISTANCE, return_default=True)
     self._follow_buttons = multiple_button_item(
       "Follow Distance",
@@ -101,6 +114,14 @@ class DrivingMannerismsLayout(Widget):
   def _on_adaptive_accel(self, state):
     self._params.put_bool(NAPParamKeys.ADAPTIVE_ACCEL, state)
 
+  def _accel_index(self, value: int) -> int:
+    if value in MAP_SPEED_ACCEL:
+      return MAP_SPEED_ACCEL.index(value)
+    return MAP_SPEED_ACCEL.index(MAP_SPEED_ACCEL_DEFAULT)
+
+  def _on_accel(self, index: int):
+    self._params.put("NAPMapSpeedAccel", MAP_SPEED_ACCEL[index])
+
   def _on_follow_distance(self, index: int):
     self._params.put(NAPParamKeys.FOLLOW_DISTANCE, index + 1)
 
@@ -115,6 +136,8 @@ class DrivingMannerismsLayout(Widget):
     self._hill_climb.action_item.set_state(self._params.get_bool(NAP_HYPERMILE_HILL_CLIMB))
     self._hill_climb.set_visible(hypermile_on)
     self._adaptive_accel.action_item.set_state(self._params.get_bool(NAPParamKeys.ADAPTIVE_ACCEL))
+    accel = int(self._params.get("NAPMapSpeedAccel", return_default=True) or MAP_SPEED_ACCEL_DEFAULT)
+    self._accel_buttons.action_item.set_selected_button(self._accel_index(accel))
     follow_dist = self._params.get(NAPParamKeys.FOLLOW_DISTANCE, return_default=True)
     self._follow_buttons.action_item.set_selected_button(max(0, min(6, follow_dist - 1)))
     self._follow_buttons.set_visible(True)